cleanup, move clock useeffect to appcontext
